Move test to file with similar tests
